feat: Add command line option for API server port
This commit adds a new command line option for specifying the port number of the API server. The `server.ts` file has been modified to use the `Command` class from the `commander` package to parse the command line arguments and set the `port` option. The server will now listen on the specified port when started.